catch an edit on a deleted message, and resend the nowplaying message… (#89)
* catch an edit on a deleted message, and resend the nowplaying message and store that

+++ src/main/java/toby/helpers/MusicPlayerHelper.java
@@ -10,6 +10,7 @@ import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEve
 import net.dv8tion.jda.api.interactions.InteractionHook;
 import net.dv8tion.jda.api.interactions.callbacks.IReplyCallback;
 import net.dv8tion.jda.api.interactions.components.buttons.Button;
+import org.springframework.web.ErrorResponseException;
 import toby.jpa.dto.MusicDto;
 import toby.jpa.dto.UserDto;
 import toby.lavaplayer.GuildMusicManager;
@@ -90,15 +91,24 @@ public class MusicPlayerHelper {
         // Get the previous "Now Playing" message if it exists
         Message previousNowPlayingMessage = guildLastNowPlayingMessage.get(guildId);
 
-        if (previousNowPlayingMessage != null) {
-            // Update the existing "Now Playing" message
-            previousNowPlayingMessage.editMessage(nowPlaying).setActionRow(pausePlay, stop).queue();
-            hook.deleteOriginal().queue();
-        } else {
+        try {
+            if (previousNowPlayingMessage != null) {
+                // Update the existing "Now Playing" message
+                previousNowPlayingMessage.editMessage(nowPlaying).setActionRow(pausePlay, stop).queue();
+                hook.deleteOriginal().queue();
+            } else {
+                sendNewNowPlayingMessage(hook, nowPlaying, pausePlay, stop, guildId);
+            }
+        } catch (IllegalArgumentException | ErrorResponseException e) {
             // Send a new "Now Playing" message and store it
-            Message nowPlayingMessage = hook.sendMessage(nowPlaying).setActionRow(pausePlay, stop).complete();
-            guildLastNowPlayingMessage.put(guildId, nowPlayingMessage);
+            sendNewNowPlayingMessage(hook, nowPlaying, pausePlay, stop, guildId);
         }
+    }
+
+    private static void sendNewNowPlayingMessage(InteractionHook hook, String nowPlaying, Button pausePlay, Button stop, long guildId) {
+        // Send a new "Now Playing" message and store it
+        Message nowPlayingMessage = hook.sendMessage(nowPlaying).setActionRow(pausePlay, stop).complete();
+        guildLastNowPlayingMessage.put(guildId, nowPlayingMessage);
     }
 
     private static boolean checkForPlayingTrack(AudioTrack track, InteractionHook hook, Integer deleteDelay) {
@@ -223,14 +233,13 @@ public class MusicPlayerHelper {
 
     private static void resetNowPlayingMessage(long guildId) {
         Message message = guildLastNowPlayingMessage.get(guildId);
-        if (message != null) {
-            message.delete().queue();
-        }
+        if (message != null) message.delete().queue();
+        guildLastNowPlayingMessage.remove(guildId);
     }
 
 
-    public static void resetMessages() {
-        guildLastNowPlayingMessage.clear();
+    public static void resetMessages(long guildId) {
+        resetNowPlayingMessage(guildId);
     }
 
 }
